training/requirements-training.txt
ultralytics==8.2.18
segment-anything==1.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
PyYAML==6.0.1
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.71.0
